Vulnerabilities reduced to 0 and extension works again
