+++ configs/dev.json
@@ -1,3 +1,4 @@
 {
-    "public": false
+    "public": false,
+    "baseUri": "https://localhost:3000"
 }
