+++ src/view.ts
@@ -3,7 +3,7 @@ import { Model } from "./model";
 export class View {
     private currentValue: string = "";
     private container!: HTMLElement;
-    private saveTimeout: number | null = null;
+    private saveTimeout: any = null;
 
     constructor(
         private model: Model,
@@ -28,200 +28,102 @@ export class View {
             document.body.appendChild(this.container);
         }
         
-        // Add a unique class to scope our CSS and prevent leakage
-        this.container.classList.add('hitcount-extension-root');
-        
-        // Use the original container styling - much more minimal but with flexbox
-        this.container.style.display = 'flex';
-        this.container.style.alignItems = 'center';
-        this.container.style.paddingLeft = '1px';
-        this.container.style.overflow = 'hidden';
-        this.container.style.overflowX = 'hidden';
-        this.container.style.overflowY = 'hidden';
-        this.container.style.maxWidth = '100%';
-        this.container.style.boxSizing = 'border-box';
-        
-        // Prevent scrolling on the entire container as well
-        this.container.addEventListener('wheel', (evt) => {
-            evt.preventDefault();
-            evt.stopPropagation();
-            return false;
-        }, { passive: false });
+        // Try to use Azure DevOps native component classes
+        this.container.className = 'bolt-textfield-container';
+        this.container.id = 'control-container';
 
-        // Create input wrapper
-        const wrap = document.createElement('div');
-        wrap.className = 'wrap combo emptyBorder';
-        wrap.style.flexGrow = '1';
+        // Create the main input using ADO classes
+        const inputGroup = document.createElement('div');
+        inputGroup.className = 'bolt-textfield';
         
-        // Prevent scrolling on the wrapper as well
-        wrap.addEventListener('wheel', (evt) => {
-            evt.preventDefault();
-            evt.stopPropagation();
-            return false;
-        }, { passive: false });
-
-        // Create number input - use text type to avoid browser number input quirks
         const input = document.createElement('input');
         input.type = 'text';
+        input.className = 'bolt-textfield-input';
         input.inputMode = 'numeric';
         input.pattern = '[0-9]*';
+        
         this.currentValue = String(this.model.getCurrentValue());
         input.value = this.currentValue;
         input.setAttribute('aria-valuenow', this.currentValue);
         
-        // More aggressive approach to disable scrolling - simplified
-        const preventScroll = (evt: Event) => {
-            evt.preventDefault();
-            evt.stopPropagation();
-            return false;
-        };
-        
-        // Disable mouse wheel scrolling on the input
-        input.addEventListener('wheel', preventScroll, { passive: false });
-        
-        // Add input validation for numbers only
+        // Add input validation and events
         input.addEventListener('input', (evt) => {
             const target = evt.target as HTMLInputElement;
-            let value = target.value;
-            
-            // Remove any non-digit characters except minus sign at the beginning
-            value = value.replace(/[^-0-9]/g, '');
-            
-            // Ensure minus sign only at the beginning
-            if (value.indexOf('-') > 0) {
-                value = value.replace(/-/g, '');
-            }
-            
-            // Update the input value
-            target.value = value;
-            
-            // Also call our debounced change handler
+            // Only allow numbers
+            target.value = target.value.replace(/[^0-9]/g, '');
             this.inputChangedWithDebounce();
         });
         
-        // Prevent non-numeric characters from being typed
         input.addEventListener('keydown', (evt) => {
-            // Allow special keys: backspace, delete, tab, escape, enter, home, end, left, right
-            if ([8, 9, 27, 13, 46, 35, 36, 37, 39].indexOf(evt.keyCode) !== -1 ||
-                // Allow Ctrl+A, Ctrl+C, Ctrl+V, Ctrl+X
-                (evt.keyCode === 65 && evt.ctrlKey) ||
-                (evt.keyCode === 67 && evt.ctrlKey) ||
-                (evt.keyCode === 86 && evt.ctrlKey) ||
-                (evt.keyCode === 88 && evt.ctrlKey)) {
-                return;
-            }
-            
-            // Allow minus sign only at the beginning
-            if (evt.key === '-' && input.selectionStart === 0) {
-                return;
-            }
-            
-            // Allow arrow keys for our custom increment/decrement
             if (evt.key === 'ArrowUp') {
                 evt.preventDefault();
                 this.handleKeyDown(evt);
-                return;
             } else if (evt.key === 'ArrowDown') {
                 evt.preventDefault();
                 this.handleKeyDown(evt);
+            }
+            // Allow: backspace, delete, tab, escape, enter
+            if ([46, 8, 9, 27, 13].indexOf(evt.keyCode) !== -1 ||
+                // Allow: Ctrl+A, Ctrl+C, Ctrl+V, Ctrl+X
+                (evt.keyCode === 65 && evt.ctrlKey === true) ||
+                (evt.keyCode === 67 && evt.ctrlKey === true) ||
+                (evt.keyCode === 86 && evt.ctrlKey === true) ||
+                (evt.keyCode === 88 && evt.ctrlKey === true)) {
                 return;
             }
-            
-            // Ensure that it is a number and stop the keypress
+            // Ensure that it's a number and stop the keypress
             if ((evt.shiftKey || (evt.keyCode < 48 || evt.keyCode > 57)) && (evt.keyCode < 96 || evt.keyCode > 105)) {
                 evt.preventDefault();
             }
         });
-
-        // Add event listeners
+        
         input.addEventListener('change', () => this.inputChanged());
-        input.addEventListener('input', () => this.inputChangedWithDebounce());
         input.addEventListener('blur', () => this.inputChanged());
+        
+        inputGroup.appendChild(input);
 
-        wrap.appendChild(input);
-
-        // Create increment button - make it bigger
-        const uptick = document.createElement('div');
-        uptick.className = 'bowtie-icon bowtie-math-plus-box';
-        uptick.style.display = 'inline-block';
-        uptick.style.cursor = 'pointer';
-        uptick.style.fontSize = '16px';
-        uptick.style.color = '#666';
-        uptick.style.padding = '6px';
-        uptick.style.userSelect = 'none';
-        uptick.style.marginLeft = '4px';
-        uptick.style.verticalAlign = 'middle';
-        uptick.title = 'Increment value';
+        // Create buttons using ADO button classes
+        const buttonGroup = document.createElement('div');
+        buttonGroup.className = 'bolt-button-group';
         
-        // Fallback text if icon doesn't load - bigger
-        if (!uptick.textContent) {
-            uptick.textContent = '+';
-            uptick.style.fontWeight = 'bold';
-            uptick.style.fontSize = '14px';
-        }
-        
-        uptick.addEventListener('click', () => {
-            if (this.onUpTick) {
-                this.onUpTick();
-            }
-        });
-
-        // Create decrement button - make it bigger
-        const downtick = document.createElement('div');
-        downtick.className = 'bowtie-icon bowtie-math-minus-box';
-        downtick.style.display = 'inline-block';
-        downtick.style.cursor = 'pointer';
-        downtick.style.fontSize = '16px';
-        downtick.style.color = '#666';
-        downtick.style.padding = '6px';
-        downtick.style.userSelect = 'none';
-        downtick.style.marginLeft = '4px';
-        downtick.style.verticalAlign = 'middle';
-        downtick.title = 'Decrement value';
-        
-        // Fallback text if icon doesn't load - bigger
-        if (!downtick.textContent) {
-            downtick.textContent = '-';
-            downtick.style.fontWeight = 'bold';
-            downtick.style.fontSize = '14px';
-        }
-        
-        downtick.addEventListener('click', () => {
+        // Decrement button
+        const decrementBtn = document.createElement('button');
+        decrementBtn.type = 'button';
+        decrementBtn.className = 'bolt-button bolt-button-subtle';
+        decrementBtn.textContent = '−';
+        decrementBtn.title = 'Decrement value';
+        decrementBtn.addEventListener('click', () => {
             if (this.onDownTick) {
                 this.onDownTick();
             }
         });
-
-        // Add hover effects
-        this.container.addEventListener('mouseenter', () => {
-            wrap.classList.add('border');
-        });
-
-        this.container.addEventListener('mouseleave', () => {
-            if (document.activeElement !== input) {
-                wrap.classList.remove('border');
+        
+        // Increment button
+        const incrementBtn = document.createElement('button');
+        incrementBtn.type = 'button';
+        incrementBtn.className = 'bolt-button bolt-button-subtle';
+        incrementBtn.textContent = '+';
+        incrementBtn.title = 'Increment value';
+        incrementBtn.addEventListener('click', () => {
+            if (this.onUpTick) {
+                this.onUpTick();
             }
         });
-
-        // Add hover effects for buttons - more subtle
-        uptick.addEventListener('mouseenter', () => {
-            uptick.style.color = '#333';
-        });
-        uptick.addEventListener('mouseleave', () => {
-            uptick.style.color = '#666';
-        });
-
-        downtick.addEventListener('mouseenter', () => {
-            downtick.style.color = '#333';
-        });
-        downtick.addEventListener('mouseleave', () => {
-            downtick.style.color = '#666';
-        });
+        
+        buttonGroup.appendChild(decrementBtn);
+        buttonGroup.appendChild(incrementBtn);
 
         // Assemble the UI
-        this.container.appendChild(wrap);
-        this.container.appendChild(downtick);
-        this.container.appendChild(uptick);
+        this.container.appendChild(inputGroup);
+        this.container.appendChild(buttonGroup);
+    }
+    
+    public update(value: number): void {
+        this.currentValue = String(value);
+        const input = this.container.querySelector('input') as HTMLInputElement;
+        if (input) {
+            input.value = this.currentValue;
+        }
     }
 
     private handleKeyDown(evt: KeyboardEvent): void {
@@ -251,20 +153,12 @@ export class View {
         if (this.saveTimeout) {
             clearTimeout(this.saveTimeout);
         }
-        
+
         // Set a new timeout to save after 500ms of no typing
         this.saveTimeout = setTimeout(() => {
             this.inputChanged();
             this.saveTimeout = null;
         }, 500);
-    }
-
-    public update(value: number): void {
-        this.currentValue = String(value);
-        const input = this.container.querySelector('input') as HTMLInputElement;
-        if (input) {
-            input.value = this.currentValue;
-        }
     }
 
     public cleanup(): void {
