--- styles/style-ado-native.css
@@ -1,145 +1,0 @@
-.bolt-textfield-container {
-    display: flex !important;
-    align-items: center !important;
-    gap: 6px !important;
-    overflow: hidden !important;
-    width: 100% !important;
-    min-width: 250px !important;
-    height: auto !important;
-    min-height: 32px !important;
-    padding: 4px !important;
-}
-
-.bolt-textfield {
-    flex: 1 !important;
-    flex-shrink: 1 !important;
-    overflow: hidden !important;
-    min-width: 120px !important;
-    max-width: 180px !important;
-    width: auto !important;
-    height: 28px !important;
-}
-
-.bolt-textfield-input {
-    overflow: hidden !important;
-    scrollbar-width: none !important;
-    -ms-overflow-style: none !important;
-    width: 100% !important;
-    height: 26px !important;
-    box-sizing: border-box !important;
-    font-size: 13px !important;
-    padding: 4px 6px !important;
-}
-
-.bolt-textfield-input::-webkit-scrollbar {
-    display: none !important;
-    width: 0 !important;
-    height: 0 !important;
-}
-
-.bolt-button-group {
-    display: flex !important;
-    gap: 3px !important;
-    overflow: hidden !important;
-    flex-shrink: 0 !important;
-    min-width: 58px !important;
-    max-width: 58px !important;
-}
-
-.bolt-button.bolt-button-subtle {
-    min-width: 26px !important;
-    max-width: 26px !important;
-    width: 26px !important;
-    height: 26px !important;
-    padding: 3px !important;
-    font-family: monospace !important;
-    font-weight: bold !important;
-    font-size: 13px !important;
-    line-height: 1 !important;
-    overflow: visible !important;
-    flex-shrink: 0 !important;
-    background-color: #f8f9fa !important;
-    border: 1px solid #ccc !important;
-    color: #333 !important;
-    cursor: pointer !important;
-    transition: all 0.2s ease !important;
-    border-radius: 3px !important;
-    display: flex !important;
-    align-items: center !important;
-    justify-content: center !important;
-}
-
-.bolt-button.bolt-button-subtle:hover {
-    background-color: #e6e6e6 !important;
-    border-color: #999 !important;
-    color: #000 !important;
-    transform: translateY(-1px) !important;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2) !important;
-}
-
-/* Hide scrollbars while allowing content to be visible */
-#control-container {
-    overflow: hidden !important;
-    scrollbar-width: none !important;
-    -ms-overflow-style: none !important;
-    width: 100% !important;
-    max-width: none !important;
-    min-width: 250px !important;
-    min-height: 32px !important;
-    box-sizing: border-box !important;
-    display: flex !important;
-    align-items: center !important;
-    gap: 6px !important;
-    padding: 4px !important;
-    flex-wrap: nowrap !important;
-}
-
-#control-container *,
-.bolt-textfield-container,
-.bolt-textfield-container * {
-    scrollbar-width: none !important;
-    -ms-overflow-style: none !important;
-    box-sizing: border-box !important;
-}
-
-#control-container::-webkit-scrollbar,
-#control-container *::-webkit-scrollbar,
-.bolt-textfield-container::-webkit-scrollbar,
-.bolt-textfield-container *::-webkit-scrollbar {
-    display: none !important;
-    width: 0 !important;
-    height: 0 !important;
-    background: transparent !important;
-}
-
-/* Target parent containers */
-.work-item-form-control-container,
-.work-item-form-control,
-.control-container {
-    overflow: hidden !important;
-    scrollbar-width: none !important;
-    -ms-overflow-style: none !important;
-    width: 100% !important;
-    max-width: none !important;
-    min-width: 250px !important;
-}
-
-.work-item-form-control-container::-webkit-scrollbar,
-.work-item-form-control::-webkit-scrollbar,
-.control-container::-webkit-scrollbar {
-    display: none !important;
-    width: 0 !important;
-    height: 0 !important;
-}
-
-/* Global scrollbar prevention */
-* {
-    scrollbar-width: none !important;
-    -ms-overflow-style: none !important;
-}
-
-*::-webkit-scrollbar {
-    display: none !important;
-    width: 0 !important;
-    height: 0 !important;
-}
